+++ demos/common/demos.js
@@ -4,7 +4,7 @@
 * DESC: Common test/demo slides for all library features
 * DEPS: Used by various demos (./demos/browser, ./demos/node, etc.)
 * VER.: 3.2.0
-* BLD.: 20200304
+* BLD.: 20200310
 */
 
 var isIE11 = typeof window !== 'undefined' && !!window['MSInputMethodContext'] && !!document['documentMode'];
@@ -2767,7 +2767,7 @@ function genSlides_Text(pptx) {
 			{ x:0.5, y:1.0, w:5.75, h:3.0, fill:'373737' }
 		);
 
-		slide.addText("Bullet Styles:", { x:0.5, y:4.5, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
+		slide.addText("Bullet Styles:", { x:0.5, y:4.6, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
 		slide.addText(
 			[
 				{ text:'style: arabicPeriod' , options:{ color:pptx.colors.ACCENT2, bullet:{type:'number', style:'arabicPeriod'} } },
@@ -2777,7 +2777,7 @@ function genSlides_Text(pptx) {
 				{ text:'style: romanLcPeriod', options:{ color:pptx.colors.ACCENT6, bullet:{type:'number', style:'romanLcPeriod'}, indentLevel:2 } },
 				{ text:'style: romanLcPeriod', options:{ color:pptx.colors.ACCENT6, bullet:{type:'number', style:'romanLcPeriod'}, indentLevel:2 } }
 			],
-			{ x:0.5, y:4.9, w:5.75, h:2.0, fill:pptx.colors.BACKGROUND2, color:pptx.colors.ACCENT1 }
+			{ x:0.5, y:5.0, w:5.75, h:2.0, fill:pptx.colors.BACKGROUND2, color:pptx.colors.ACCENT1 }
 		);
 
 		// RIGHT COLUMN ------------------------------------------------------------
@@ -2786,33 +2786,41 @@ function genSlides_Text(pptx) {
 		slide.addText("Bullet startAt number option:", { x:7.0, y:0.6, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
 		slide.addText(
 			"type:'number'\nstartAt:'5'",
-			{ x:7.0, y:1.05, w:'40%', h:0.75, fill:pptx.colors.BACKGROUND2, color:pptx.colors.ACCENT6, fontFace:"Courier New", bullet:{type:'number', startAt:'5'} }
+			{ x:7.0, y:1.0, w:'40%', h:0.75, fill:pptx.colors.BACKGROUND2, color:pptx.colors.ACCENT6, fontFace:"Courier New", bullet:{type:'number', startAt:'5'} }
 		);
 
 		// 2: Bullets: Text With Line-Breaks
-		slide.addText("Bullets made with Line Breaks:", { x:7.0, y:2.1, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
-		slide.addText('Line 1\nLine 2\nLine 3', { x:7.0, y:2.6, w:'40%', h:1.0, color:'393939', fontSize:16, fill:'F2F2F2', bullet:{type:'number'} });
+		slide.addText("Bullets made with Line Breaks:", { x:7.0, y:2.0, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
+		slide.addText('Line 1\nLine 2\nLine 3', { x:7.0, y:2.4, w:'40%', h:1.0, color:'393939', fontSize:16, fill:'F2F2F2', bullet:{type:'number'} });
 
 		// 3: Bullets: With group of {text}
-		slide.addText("Bullets made with text objects:", { x:7.0, y:4.0, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
+		slide.addText("Bullets with marginPt:", { x:7.0, y:3.65, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
 		slide.addText(
 			[
-				{ text: 'big red words... ', options:{fontSize:24, color:'FF0000'} },
-				{ text: 'some green words.', options:{fontSize:16, color:'00FF00'} }
+				{ text: 'bullet marginPt:10', options: { bullet: { marginPt: 10 } } },
+				{ text: 'bullet marginPt:30', options: { bullet: { marginPt: 30 } } }
 			],
-			{ x:7.0, y:4.4, w:5.75, h:0.5, margin:0.1, fontFace:'Arial', bullet:{code:'25BA'} }
+			{ x:7.0, y:3.95, w:5.75, h:0.5, margin:0.1, fontFace:'Arial', fontSize:12 }
+		);
+
+		slide.addText("Bullets with text objects:", { x:7.0, y:4.7, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
+		slide.addText(
+			[
+				{ text: 'big red words... ', options: { fontSize: 24, color: 'FF0000' } },
+				{ text: 'some green words.', options: { fontSize: 16, color: '00FF00' } }
+			],
+			{ x: 7.0, y: 5.0, w: 5.75, h: 0.4, margin: 0.1, fontFace: 'Arial', bullet: { code: '25BA' } }
 		);
 
 		// 5: Bullets: Within a {text} object
-		slide.addText("Bullets using text object options:", { x:7.0, y:5.3, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
+		slide.addText("Bullets using text object options:", { x:7.0, y:5.65, w:'40%', h:0.3, margin:0, color:pptx.colors.ACCENT1 });
 		slide.addText(
 			[
-				{ text:'I am a text object with bullets..', options:{bullet:{code:'25BA'}, color:'CC0000'} },
-				{ text:'and I am the next text object.'   , options:{bullet:{code:'25BA'}, color:'00CD00'} },
-				{ text:'Default bullet text.. '           , options:{bullet:true, color:'696969'} },
-				{ text:'Final text object w/ bullet:true.', options:{bullet:true, color:'0000AB'} }
+				{ text: 'bullet:{code:"25BA"}, color:"00CD00"', options: { bullet: { code: '25BA' }, color: '00CD00' } },
+				{ text: 'Next bullet text', options: { bullet: true, color: '696969' } },
+				{ text: 'LAst bullet text', options: { bullet: true, color: '0000AB' } }
 			],
-			{ x:7.0, y:5.65, w:'40%', h:1.4, color:'ABABAB', margin:1 }
+			{ x: 7.0, y: 5.95, w: '40%', h: 1.05, color: 'ABABAB', margin: 1 }
 		);
 	}